Open the output writer only after the users responds to the prompt.
In the case that the user does a git rebase on their log file after
starting the program but before responding to the prompt, the file
handle would have been clobbered and the report would have been lost.
This may or may not have happened to the author of this commit.